VoxelEngine porting progress

+++ Source/Engine/ReCrafted.Engine/Voxels/Meshing/Transvoxel/TransvoxelMesher.h
@@ -13,24 +13,27 @@
 
 #include <bitset>
 
+/// <summary>
+///     The voxel mesher implementation class of Transvoxel algorithm.
+/// </summary>
 class TransvoxelMesher final : public IVoxelMesher
 {
     DELETE_COPY_MOVE(TransvoxelMesher);
 
 private:
-    /**
-     * \brief The maximal amount of materials per mesh section.
-     */
+    /// <summary>
+    ///     The maximal amount of materials per mesh section.
+    /// </summary>
     static const int MaxMaterialsPerSection = 16;
 
-    /**
-    * \brief The maximal amount of mesh section.
-    */
+    /// <summary>
+    ///     The maximal amount of mesh section.
+    /// </summary>
     static const int MaxSections = 8;
 
-    /**
-    * \brief Enables or disables transition cell generation.
-    */
+    /// <summary>
+    ///     Enables or disables transition cell generation.
+    /// </summary>
     static const bool EnableTransitionCells = false;
 
     static const int VertexReuseEntriesCount = (VoxelChunkData::ChunkDataSize * 3) * (VoxelChunkData::ChunkDataSize * 3) * (VoxelChunkData::ChunkDataSize * 3);
@@ -86,7 +89,7 @@ private:
     public:
         void AddVertex(const VertexInfo& vertexInfo, bool normalCorrection)
         {
-            auto index = vertices.Count();
+            auto index = static_cast<uint>(vertices.Count());
 
             if (!vertexReuseMap.test(vertexInfo.vertexId))
             {
@@ -186,21 +189,22 @@ private:
             return MaxMaterialsPerSection - materialSet.Count() >= neededEntries;
         }
 
-        /**
-        * \brief Returns true when this material set already contains given material.
-        * \param material The material.
-        */
+        /// <summary>
+        ///     Returns true when this material set already contains given material.
+        /// </summary>
+        /// <param name="material">The material.</param>
         bool HasMaterial(const VoxelMaterial_t material) const
         {
             return materialSet.Contains(material);
         }
 
-        /**
-        * \brief Returns true when this material set already contains given material.
-        * \param materialA The first material.
-        * \param materialB The second material.
-        * \param materialC The third material.
-        */
+        /// <summary>
+        ///     Returns true when this material set already contains given material.
+        /// </summary>
+        /// <param name="materialA">The first material.</param>
+        /// <param name="materialB">The second material.</param>
+        /// <param name="materialC">The third material.</param>
+        /// <returns></returns>
         bool HasMaterials(const VoxelMaterial_t materialA, const VoxelMaterial_t materialB, const VoxelMaterial_t materialC) const
         {
             if (materialA == materialB && materialB == materialC)
@@ -209,6 +213,9 @@ private:
             return HasMaterial(materialA) && HasMaterial(materialB) && HasMaterial(materialC);
         }
 
+        /// <summary>
+        ///     Clears all cached data.
+        /// </summary>
         void Clear()
         {
             // Clear all arrays
@@ -284,39 +291,40 @@ private:
     void PolygonizeRegularCell(const Vector3& position, Voxel* data, float voxelScale, int lod, const Vector3i& voxelOffset, bool normalCorrection);
 
 public:
-    /**
-    * \brief Virtual method for generating a mesh from hermite voxel data.
-    * \param position The position of current procesed chunk.
-    * \param lod The current procesed chunk lod.
-    * \param borders Border set for current procesed chunk.
-    * \param data The hermite voxel data (in -127 to 127 range).
-    */
+    /// <summary>
+    ///     Virtual method for generating a mesh from hermite voxel data.
+    /// </summary>
+    /// <param name="position">The position of current processed chunk.</param>
+    /// <param name="lod">The current processed chunk lod.</param>
+    /// <param name="borders">Border set for current processed chunk.</param>
+    /// <param name="data">The hermite voxel data (in -127 to 127 range).</param>
     void Generate(const Vector3& position, int lod, uint8_t borders, Voxel* data) override;
 
-    /**
-    * \brief Checks if there are any generated triangles.
-    * \return True when there is at least one triangle.
-    */
+    /// <summary>
+    ///     Checks if there are any generated triangles.
+    /// </summary>
+    /// <returns>True when there is at least one triangle.</returns>
     bool HasTriangles() override
     {
         return m_triangleCount > 0u;
     }
 
-    /**
-    * \brief Uploads all data to a mesh. This also clears the mesher and prepares to next mesh generation.
-    * \param chunkMesh The mesh that will get the new mesh data.
-    */
+    /// <summary>
+    ///     Uploads all data to a mesh. This also clears the mesher and prepares to next mesh generation.
+    /// </summary>
+    /// <param name="chunkMesh">The mesh that will get the new mesh data.</param>
+    /// <param name="chunkCollision">The collision that will get the new collision data.</param>
     void Apply(const RefPtr<VoxelChunkMesh>& chunkMesh, RefPtr<VoxelChunkCollision>& chunkCollision) override;
 
-    /**
-    * \brief Cleans all data used during Generate and Apply functions.
-    */
+    /// <summary>
+    ///     Cleans all data used during Generate and Apply functions.
+    /// </summary>
     void Clear() override;
 
 public:
-    /**
-     * \brief Gets the name of mesher algorithm.
-     */
+    /// <summary>
+    ///     Gets the name of mesher algorithm.  
+    /// </summary>
     const char* GetName() override
     {
         return "Transvoxel";
